+++ src/main.tsx
@@ -3,6 +3,9 @@ import { flushSync } from "react-dom";
 import { createRoot } from "react-dom/client";
 import App from "./App";
 import { initAppearance, revealWindow } from "./appearance";
+import "@astryxdesign/core/reset.css";
+import "@astryxdesign/core/astryx.css";
+import "@astryxdesign/theme-neutral/theme.css";
 import "./index.css";
 
 void (async () => {